Fix to non existent git switch home folder, and minor fixes

+++ gitswitch/gitswitch.py
@@ -21,17 +21,30 @@ def handler():
   if args.backup:
     backup_ssh()
 
+def create_switch_home():
+  home_location = os.environ.get('HOME')
+  switch_location = os.path.join(home_location,'.git-switch')
+  os.makedirs(switch_location, 0755)
+
+
 def get_switch_location():
   home_location = os.environ.get('HOME')
   return os.path.join(home_location,'.git-switch')
   
 def list_environments():
-  directories =  os.listdir(get_switch_location())
-  print "Current environments found:"
-  for directory in directories:
-    current_directory_path = os.path.join(get_switch_location(), directory)
-    if os.path.isdir(current_directory_path):
-      print "* {}".format(directory)
+  if os.path.exists(get_switch_location()):
+    directories =  os.listdir(get_switch_location())
+    if len(directories) == 0:
+      print "You need to add some folders to $HOME/.git-switch"
+      return
+    print "Current environments found:"
+    for directory in directories:
+      current_directory_path = os.path.join(get_switch_location(), directory)
+      if os.path.isdir(current_directory_path):
+        print "* {}".format(directory)
+  else:
+    create_switch_home()
+    print "This is the first time you are running git-switch, please add some folders to $HOME/.git-switch"
 
 def activate_environment(environment):
   print "Activate environment"
@@ -40,8 +53,10 @@ def activate_environment(environment):
   ssh_folder = os.path.join(os.environ.get('HOME'),'.ssh/')
   if os.path.exists(ssh_folder):
     shutil.rmtree(ssh_folder)
-  shutil.copytree(environment_folder, os.path.join(os.environ.get('HOME'),'.ssh'))
-
+  if os.path.exists(environment_folder):
+    shutil.copytree(environment_folder, os.path.join(os.environ.get('HOME'),'.ssh'))
+  else:
+    print "The enviroment you are trying to activate does not exists"
 
 def create_environment():
   print "Create environment"
